# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Business API",
    "version": "2.0.0",
    "description": "Enterprise-grade middleware API for CAC.gov.ng business registration. All 46 core routes documented."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Use 'Authorization: Token ck_...' for customer API key auth"
      },
      "CustomerApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Customer API key in format: Token ck_your_api_key_here"
      }
    },
...
  },
  "paths": {
...
    "/admin/admins": {
      "post": {
        "description": "Create a new admin user (super admin only)",
        "tags": [
          "Admin - System Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "properties": {
                        "admin": {
                          "properties": {
                            "email": {
                              "type": "string"
                            },
                            "permissions": {
                              "items": {
                                "type": "string"
                              },
                              "type": "array"
                            },
                            "id": {
                              "type": "string"
                            },
                            "role": {
                              "type": "string"
                            }
                          },
                          "type": "object"
                        }
                      },
                      "type": "object"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Admin created successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          }
        },
        "summary": "Create new admin user",
        "operationId": "createAdmin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "email": {
                    "format": "email",
                    "example": "newadmin@company.com",
                    "type": "string"
                  },
                  "permissions": {
                    "example": [
                      "view_all",
                      "manage_customers"
                    ],
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "password": {
                    "type": "string",
                    "example": "AdminPass123"
                  },
                  "role": {
                    "enum": [
                      "admin",
                      "super_admin"
                    ],
                    "example": "admin",
                    "type": "string"
                  }
                },
                "type": "object",
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          },
          "required": true
        }
      }
    },
...
  }
}